fix: insert alert file into alert package
